requirements.txt
# ── Core Agents ────────────────────────────────────────────────
# crewai 0.86+ ships its own langchain pins; let pip resolve sub-deps
crewai>=0.86.0,<1.0.0
crewai-tools>=0.20.0,<1.0.0

# ── LangChain ecosystem ─────────────────────────────────────────
# crewai 0.86 requires langchain >=0.3.0; chromadb 0.5+ works with these
langchain>=0.3.14,<0.4.0
langchain-community>=0.3.14,<0.4.0
langchain-groq>=0.2.3,<0.3.0
langchain-huggingface>=0.1.2,<0.2.0

# ── Vector Store ────────────────────────────────────────────────
# chromadb 0.5.x has stable wheels on Python 3.10/3.11 and avoids
# the grpcio conflicts introduced in 0.6.x
chromadb>=0.5.0,<0.6.0

# ── Embeddings ──────────────────────────────────────────────────
sentence-transformers>=2.7.0,<4.0.0

# ── PDF Processing ──────────────────────────────────────────────
pypdf>=4.0.0,<6.0.0
pdfplumber>=0.10.0,<0.12.0

# ── Data ────────────────────────────────────────────────────────
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0

# ── UI ──────────────────────────────────────────────────────────
streamlit>=1.35.0,<2.0.0

# ── Utilities ───────────────────────────────────────────────────
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.0,<7.0.0
tiktoken>=0.7.0,<1.0.0
tenacity>=8.2.0,<10.0.0
rich>=13.0.0,<14.0.0

# ── Groq ────────────────────────────────────────────────────────
groq>=0.11.0,<1.0.0
